Allows the usage of vim keys (hjkl) to move the snake
Some of the users, such as me, prefer to use the classical "vim keys"
( h - left; j - down; k - top; l - right ) instead of the popular WASD,
currently used in this project. So, I think it's a good idea to
allow the possibility to move the snake around using the vim keys since
these keys aren't "special keys" in the game.

To implement this feature the switch case used to identify the letters
readded as input from the user's keyboard was modified to detect "h j k l"
as the same as "a s w d " respectively. Also, the main menu was modified
to inform the user that now it can use the marvelous vim keys to play
the game.