Fix blank border in the play scene
The bug was fixed simply changing the
snake initial position.

Since the border of the game is drawn
only once and at the start of the game,
the blank line between elements of the
body of the snake could erase permanently
parts of the border if at the start the game
snake is close to the border, and his body
passes through the border.